backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.2
pydantic[email]>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
scikit-learn>=1.6.0
numpy>=2.0.0
pandas>=2.2.0
joblib>=1.3.2
websockets>=12.0
bcrypt>=4.1.1
